hackfiniti/notibuzz/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pinecone-client==2.2.4
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1
asyncio==3.4.3
aiofiles==23.2.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sentence-transformers==2.2.2
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.36.0
torch==2.1.1
